v5.13.26

+++ app/Repositories/BankIntegrationRepository.php
@@ -44,7 +44,7 @@ class BankIntegrationRepository extends BaseRepository
             return $bank_integration;
         }
 
-        if (Ninja::isHosted()) {
+        if (Ninja::isHosted() && $bank_integration->integration_type === BankIntegration::INTEGRATION_TYPE_YODLEE) {
 
             $account = $bank_integration->account;
 
@@ -52,10 +52,21 @@ class BankIntegrationRepository extends BaseRepository
 
             $yodlee = new Yodlee($bank_integration_account_id);
 
-            try {
-                $yodlee->deleteAccount($bank_integration->bank_account_id);
-            } catch (\Exception $e) {
+            $still_referenced = BankIntegration::query()
+                                                ->where('account_id', $bank_integration->account_id)
+                                                ->where('bank_account_id', $bank_integration->bank_account_id)
+                                                ->where('id', '!=', $bank_integration->id)
+                                                ->where('is_deleted', 0)
+                                                ->exists();
 
+            if(!$still_referenced) {
+                try {
+                    $yodlee->deleteAccount($bank_integration->bank_account_id);
+                } catch (\Throwable $e) {
+
+                    nlog("YODLEE: DELETE: {$e->getMessage()}");
+                    return $bank_integration;
+                }
             }
 
         }
